refactor: update paths alias
It didn't work until now so i just reset my laptop and something happened
