+++ README.md
@@ -3,6 +3,5 @@
 - [Tasks](TODO.md)
 
 #### RemainingTasks
-1. Refactor paths alias
-2. Improve async operations via custom hook
-3. Add test cases
+1. Improve async operations via custom hook
+2. Add test cases
