+++ src/App.tsx
@@ -4,8 +4,8 @@ import Pages from "./pages/Pages";
 import FakeLogin from "pages/login/FakeLogin";
 
 function App() {
-  const { isLoginned } = useAuth();
-  return isLoginned ? <Pages /> : <FakeLogin />;
+  const { isLogin } = useAuth();
+  return isLogin ? <Pages /> : <FakeLogin />;
 }
 
 export default App;
