+++ src/index.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from "react-dom/client";
 
 import App from "./App";
-import AppProviders from "./contexts/AppProviders";
+import AppProviders from "contexts/AppProviders";
 import reportWebVitals from "./reportWebVitals";
 
 const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
