+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+import Detail from "./detail/Detail";
+import Main from "./main/Main";
+
+export { Detail, Main };
