+++ src/pages/main/components/index.ts
@@ -1,0 +1,4 @@
+import Filter from "./Filter";
+import Search from "./Search";
+
+export { Filter, Search };
